refactor `models` in `admin.branding`
