+++ features/admin.branding.v1/api/delete-custom-text-preference.ts
@@ -23,8 +23,8 @@ import { AxiosError, AxiosRequestConfig, AxiosResponse } from "axios";
 import { I18nConstants } from "../../admin.core.v1/constants/i18n-constants";
 import { store } from "../../admin.core.v1/store";
 import { OrganizationType } from "../../admin.organizations.v1/constants/organization-constants";
+import { BrandingPreferenceTypes } from "../../common.branding.v1/models/branding-preferences";
 import { CustomTextPreferenceConstants } from "../constants/custom-text-preference-constants";
-import { BrandingPreferenceTypes } from "../models/branding-preferences";
 import { CustomTextPreferenceAPIResponseInterface } from "../models/custom-text-preference";
 
 /**
