+++ features/admin.branding.v1/api/use-get-custom-text-preference-fallbacks.ts
@@ -24,7 +24,7 @@ import useRequest, {
     RequestErrorInterface,
     RequestResultInterface
 } from "../../admin.core.v1/hooks/use-request";
-import { BrandingPreferenceTypes } from "../models/branding-preferences";
+import { BrandingPreferenceTypes } from "../../common.branding.v1/models/branding-preferences";
 import { CustomTextPreferenceAPIResponseInterface } from "../models/custom-text-preference";
 
 /**
