+++ features/admin.branding.v1/components/branding-core.tsx
@@ -43,6 +43,13 @@ import { AppState } from "../../admin.core.v1/store";
 import { ExtendedFeatureConfigInterface } from "../../admin.extensions.v1/configs/models";
 import { useGetCurrentOrganizationType } from "../../admin.organizations.v1/hooks/use-get-organization-type";
 import { OrganizationResponseInterface } from "../../admin.organizations.v1/models/organizations";
+import {
+    BrandingPreferenceAPIResponseInterface,
+    BrandingPreferenceInterface,
+    BrandingPreferenceLayoutInterface,
+    BrandingPreferenceThemeInterface,
+    PredefinedLayouts
+} from "../../common.branding.v1/models";
 import { deleteBrandingPreference, updateBrandingPreference } from "../api";
 import deleteAllCustomTextPreferences from "../api/delete-all-custom-text-preference";
 import useGetBrandingPreferenceResolve from "../api/use-get-branding-preference-resolve";
@@ -51,13 +58,6 @@ import { BrandingPreferenceTabs, DesignFormValuesInterface } from "../components
 import { BrandingPreferencesConstants } from "../constants";
 import { CustomTextPreferenceConstants } from "../constants/custom-text-preference-constants";
 import { BrandingPreferenceMeta, LAYOUT_PROPERTY_KEYS } from "../meta";
-import {
-    BrandingPreferenceAPIResponseInterface,
-    BrandingPreferenceInterface,
-    BrandingPreferenceLayoutInterface,
-    BrandingPreferenceThemeInterface,
-    PredefinedLayouts
-} from "../models";
 import { BrandingPreferenceUtils } from "../utils";
 
 /**
