+++ features/admin.branding.v1/components/branding-preference-tabs.tsx
@@ -49,17 +49,17 @@ import ScreenDropdown from "./screen-dropdown";
 import { StickyTabPaneActionPanel } from "./sticky-tab-pane-action-panel";
 import { commonConfig } from "../../admin.extensions.v1/configs";
 import { AppState } from "../../admin.core.v1/store";
-import { BrandingPreferencesConstants } from "../constants";
-import { CustomTextPreferenceConstants } from "../constants/custom-text-preference-constants";
-import useBrandingPreference from "../hooks/use-branding-preference";
-import { BrandingPreferenceMeta } from "../meta";
 import {
     BrandingPreferenceInterface,
     BrandingPreferenceThemeInterface,
     BrandingSubFeatures,
     PredefinedThemes,
     PreviewScreenType
-} from "../models";
+} from "../../common.branding.v1/models";
+import { BrandingPreferencesConstants } from "../constants";
+import { CustomTextPreferenceConstants } from "../constants/custom-text-preference-constants";
+import useBrandingPreference from "../hooks/use-branding-preference";
+import { BrandingPreferenceMeta } from "../meta";
 import { CustomTextConfigurationModes } from "../models/custom-text-preference";
 
 /**
