+++ features/admin.branding.v1/components/custom-text/custom-text.tsx
@@ -33,9 +33,9 @@ import React, {
 import { useTranslation } from "react-i18next";
 import CustomTextFields from "./custom-text-fields";
 import CustomTextUnsavedChangesConfirmationModal from "./custom-text-unsaved-changes-confirmation-modal";
+import { BrandingSubFeatures, PreviewScreenType } from "../../../common.branding.v1/models";
 import { CustomTextPreferenceConstants } from "../../constants/custom-text-preference-constants";
 import useBrandingPreference from "../../hooks/use-branding-preference";
-import { BrandingSubFeatures, PreviewScreenType } from "../../models";
 import { CustomTextConfigurationModes, CustomTextInterface } from "../../models/custom-text-preference";
 import replaceObjectKeySymbols from "../../utils/replace-object-key-symbols";
 import LocaleDropdown from "../locale-dropdown";
