+++ features/admin.branding.v1/components/design/illustrations-preview.tsx
@@ -17,7 +17,7 @@
  */
 
 import React, { ReactElement } from "react";
-import { BrandingPreferenceColorsInterface } from "../../models/branding-preferences";
+import { BrandingPreferenceColorsInterface } from "../../../common.branding.v1/models/branding-preferences";
 
 /* eslint-disable max-len */
