+++ features/admin.branding.v1/components/design/layout-swatch/layout-swatch.tsx
@@ -20,7 +20,7 @@ import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import classNames from "classnames";
 import React, { FunctionComponent, MouseEventHandler, PropsWithChildren, ReactElement, useState } from "react";
 import { Placeholder } from "semantic-ui-react";
-import { BrandingPreferenceImageInterface } from "../../../models";
+import { BrandingPreferenceImageInterface } from "../../../../common.branding.v1/models";
 
 /**
  * Proptypes for the Layout Swatch component.
