+++ features/admin.branding.v1/components/preview/email-template-screen-skeleton.tsx
@@ -23,10 +23,10 @@ import Mustache from "mustache";
 import React, { FunctionComponent, ReactElement, useMemo } from "react";
 import { useSelector } from "react-redux";
 import { AppState } from "../../../admin.core.v1/store";
+import { BrandingPreferenceInterface, PredefinedThemes } from "../../../common.branding.v1/models";
 import { CustomTextPreferenceConstants } from "../../constants/custom-text-preference-constants";
 import useBrandingPreference from "../../hooks/use-branding-preference";
 import { BrandingPreferenceMeta } from "../../meta/branding-preference-meta";
-import { BrandingPreferenceInterface, PredefinedThemes } from "../../models";
 
 /**
  * Prop-types for the login box component of login screen skeleton.
