+++ features/admin.branding.v1/components/preview/product-footer.tsx
@@ -29,7 +29,7 @@ import { BrandingPreferencesConstants } from "../../constants";
 import { CustomTextPreferenceConstants } from "../../constants/custom-text-preference-constants";
 import useBrandingPreference from "../../hooks/use-branding-preference";
 import { BrandingPreferenceMeta } from "../../meta";
-import { BrandingPreferenceInterface } from "../../models";
+import { BrandingPreferenceInterface } from "../../../common.branding.v1/models";
 
 /**
  * Proptypes for the product footer component of login screen skeleton.
