+++ features/admin.branding.v1/constants/branding-preferences-constants.ts
@@ -18,8 +18,8 @@
 
 import { IdentityAppsError } from "@wso2is/core/errors";
 import { commonConfig } from "../../admin.extensions.v1";
+import { BrandingPreferenceInterface, PredefinedThemes, ThemeConfigInterface } from "../../common.branding.v1/models";
 import { PredefinedLayouts } from "../meta/layouts";
-import { BrandingPreferenceInterface, PredefinedThemes, ThemeConfigInterface } from "../models";
 
 /**
  * Class containing Branding Preferences constants.
