+++ features/admin.branding.v1/meta/branding-preference-meta.ts
@@ -21,8 +21,6 @@ import { StringUtils } from "@wso2is/core/utils";
 import isEmpty from "lodash-es/isEmpty";
 import { LAYOUTS } from "./layouts";
 import { THEMES, THEME_SWATCH_UI_CONFIGS } from "./themes";
-import { ThemeSwatchUIConfigsInterface } from "../components";
-import { BrandingPreferencesConstants } from "../constants";
 import {
     BrandingPreferenceImagesInterface,
     BrandingPreferenceOrganizationDetailsInterface,
@@ -30,7 +28,9 @@ import {
     DynamicBrandingPreferenceLayoutInterface,
     DynamicBrandingPreferenceThemeInterface,
     PredefinedThemes
-} from "../models";
+} from "../../common.branding.v1/models";
+import { ThemeSwatchUIConfigsInterface } from "../components";
+import { BrandingPreferencesConstants } from "../constants";
 
 /**
  * Class for the Branding Preference Metadata.
